Extract test blocks in files
